feat: concurrent mining

+++ internal/core/genesis.go
@@ -8,7 +8,7 @@ type Genesis struct {
 	Date                time.Time         `json:"date"`
 	ChainID             uint64            `json:"chainID"`
 	TransactionPerBlock uint64            `json:"transactionPerBlock"`
-	Difficulty          uint64            `json:"difficulty"`
+	Difficulty          uint16            `json:"difficulty"`
 	MiningReward        uint64            `json:"miningReward"`
 	GasPrice            uint64            `json:"gasPrice"`
 	Balances            map[string]uint64 `json:"balances"`
